requirements.txt
Django>=4.2
djangorestframework
stripe
firebase-admin
python-dotenv